pyproject.toml
[project]
name = "counterfactual-clinical-xai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic==0.25.0",
    "datasets==2.21.0",
    "diskcache==5.6.0",
    "groq==0.5.0",
    "httpx==0.27.2",
    "jsonlines==4.0.0",
    "litellm==1.52.0",
    "matplotlib==3.8.0",
    "numpy==1.26.4",
    "openai==1.54.4",
    "pandas==2.2.0",
    "python-dotenv==1.0.1",
    "pyyaml==6.0.2",
    "scikit-learn==1.4.0",
    "scipy==1.12.0",
    "seaborn==0.13.0",
    "sentence-transformers==2.7.0",
    "tenacity==8.2.0",
    "tqdm==4.66.5",
]
